Add files via upload
A NetLogo program that simulates MTB swimming and their events. The NetLogo programs has a few different built-in ways of tweaking its simulation, such as adjusting event probabilities, visualizing the events, built-in noise, and the length of the events. There is also an option to save the positional data and the visual data.